+++ Changelog.md
@@ -4,6 +4,7 @@ OpenCore Changelog
 #### v0.5.1
 - Added support of kernel resource kext injection
 - Added support for 0.25% clock slowdown on Xeon Scalable CPUs (thx @mrmiller)
+- Replaced `MatchKernel` with `MinKernel` and `MaxKernel`
 
 #### v0.5.0
 - Added builtin firmware versions for new models 2019